E2E-Miner/requirements.txt
#annotated-types==0.7.0
#anyio==3.6.2
#apturl==0.5.2
#asgiref==3.5.0
#async-timeout==4.0.3
#asyncpg==0.29.0
#attrs==21.2.0
#bcrypt==3.2.0
#blinker==1.4
#Brlapi==0.8.3
#certifi==2020.6.20
#cffi==1.15.1
#chardet==4.0.0
#charset-normalizer==3.3.2
#click==8.1.3
#colorama==0.4.4
#command-not-found==0.3
#contourpy==1.3.0
#cryptography==3.4.8
#cupshelpers==1.0
#cycler==0.12.1
#dbus-python==1.2.18
#defer==1.0.6
#distro==1.7.0
#distro-info==1.1+ubuntu0.2
#dnspython==2.6.1
#docker==5.0.3
#docker-compose==1.29.2
#dockerpty==0.4.1
#docopt==0.6.2
#duplicity==0.8.21
#email_validator==2.2.0
#et-xmlfile==1.1.0
#exceptiongroup==1.1.1
#fastapi==0.95.0
#fastapi-cli==0.0.4
#fasteners==0.14.1
#fonttools==4.54.1
#future==0.18.2
#gitdb==4.0.11
GitPython==3.1.43
#greenlet==2.0.2
#gyp==0.1
#h11==0.14.0
#httpcore==1.0.5
#httplib2==0.20.2
#httptools==0.6.1
#httpx==0.27.0
#idna==3.4
#importlib-metadata==4.6.4
#inflection==0.5.1
#iniconfig==2.0.0
#javalang @ git+https://github.com/jose/javalang.git@3bf4bf8681e25f7126ea2dca863335625cea312e
#jeepney==0.7.1
#jsonschema==3.2.0
#keyring==23.5.0
#kiwisolver==1.4.7
#language-selector==0.1
#launchpadlib==1.10.16
#lazr.restfulclient==0.14.4
#lazr.uri==1.0.6
#lizard==1.17.10
#lockfile==0.12.2
#louis==3.20.0
#macaroonbakery==1.3.1
#Mako==1.1.3
#mando==0.7.1
#markdown-it-py==3.0.0
#MarkupSafe==2.1.2
#matplotlib==3.9.2
#mdurl==0.1.2
#mercurial==6.1.1
#monotonic==1.6
#more-itertools==8.10.0
#netifaces==0.11.0
#numpy==2.1.1
#oauthlib==3.2.0
#olefile==0.46
#openai==1.35.14
openpyxl==3.1.5
#orjson==3.10.6
#osmium==3.2.0
#packaging==23.1
pandas==2.2.3
#paramiko==2.9.3
#pexpect==4.8.0
#Pillow==9.0.1
#pluggy==1.0.0
#protobuf==3.12.4
#psycopg2==2.9.2
#psycopg2-binary==2.9.5
#psycopg2cffi==2.9.0
#ptyprocess==0.7.0
#pycairo==1.20.1
#pycparser==2.21
#pycups==2.0.1
#pydantic==1.10.7
#pydantic_core==2.14.6
PyDriller==2.6
#Pygments==2.18.0
#PyGObject==3.42.1
#PyJWT==2.3.0
#pymacaroons==0.13.0
#PyNaCl==1.5.0
#pyparsing==2.4.7
#pyperclip==1.8.2
#pyRFC3339==1.1
#pyrsistent==0.18.1
#pytest==7.3.1
#python-apt==2.4.0+ubuntu4
#python-dateutil==2.9.0.post0
#python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
#python-multipart==0.0.9
#pytz==2022.1
#pyxdg==0.27
#PyYAML==5.4.1
#radon==6.0.1
#reportlab==3.6.8
requests==2.31.0
#rich==13.7.1
#SecretStorage==3.3.1
#selenium==3.141.0
#shellingham==1.5.4
#six==1.16.0
#smmap==5.0.1
#sniffio==1.3.0
SQLAlchemy==1.4.41
#sqlalchemy2-stubs==0.0.2a32
#sqlmodel==0.0.8
#ssh-import-id==5.11
#starlette==0.26.1
#systemd-python==234
#texttable==1.6.4
#tomli==2.0.1
#tqdm==4.66.4
#typer==0.12.3
#types-pytz==2024.2.0.20241003
#typing_extensions==4.12.2
#tzdata==2024.2
#ubuntu-drivers-common==0.0.0
#ubuntu-pro-client==8001
#ufw==0.36.1
#ujson==5.10.0
#unattended-upgrades==0.1
#urllib3==1.26.5
#usb-creator==0.3.7
#uvicorn==0.20.0
#uvloop==0.19.0
#vboxapi==1.0
#wadllib==1.3.6
#watchfiles==0.22.0
#websocket-client==1.2.3
#websockets==12.0
#Werkzeug==2.2.3
#wsproto==1.0.0
#xdg==5
#xkit==0.0.0
#zipp==1.0.0